pyproject.toml
[project]
name = "minio-operator-ext"
version = "0.0.0"
dependencies = [
    "click",
    "minio",
    "bfiola-operator-core==2.0.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "bfiola-devtools"
]

[project.scripts]
minio-operator-ext = "minio_operator_ext.cli:main"

[tool.semantic_release.branches.dev]
match = "dev"
prerelease = true
prerelease_token = "rc"

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.setuptools.packages.find]
include = ["minio_operator_ext"]
